Potential fix for code scanning alert no. 13: Multiplication result converted to larger type

+++ src/molecule.c
@@ -1850,7 +1850,7 @@ static int TestBuriedSurface( RAtom __far *aptr, RAtom __far *dptr, RAtom __far 
      */
 	
 	decun = (Long)(rint(sqrt((double)(ECxUn[0]*ECxUn[0] + ECxUn[1]*ECxUn[1] + ECxUn[2]*ECxUn[2]))));
-	dist += crad*crad + 2*(crad*decun)/4096;
+	dist += ((Long)crad)*crad + 2*(((Long)crad)*decun)/4096;
 	if (dist < erad*erad) return 0;
     
     return 1;
